Final 1 excersize

+++ cw01/zad1/list.c
@@ -40,7 +40,7 @@ void List_push(List *list, struct info *value)
 
 }
 
-void List_add(List * list, const char * firstName, const char * lastName, const char * birthDate, const char * email, int phone, const char * address) {
+void List_add(List * list, const char * firstName, const char * lastName, const char * birthDate, const char * email, long phone, const char * address) {
   struct info data;
   data.firstName = firstName;
   data.lastName = lastName;
